connect to IPv6 servers correctly without -6 option, fix printing IPv6 URLs
bind_socket was defaulting to AF_INET if no protocol family was specified.
We now do the getaddrinfo in connect_socket *before* calling bind_socket,
so that there's a chance of getting gai_hints.ai_family == AF_INET6.

Also, IPv6 URLs are printed as [::1]:80.  Do this in get_sock_name.


git-archimport-id: bonzini@gnu.org--2004b/netcat--devo--2.0--patch-6